requirements.txt
openai
pandas
numpy
python-dotenv
altair
googletrans-py
scikit-learn
streamlit
plotly